+++ CHANGELOG.md
@@ -16,6 +16,10 @@ x.x.x Release notes (yyyy-MM-dd)
 * Fix an issue where values set on a Realm object using `setValue(value:, forKey:)`
   that were not themselves Realm objects were not properly converted into Realm
   objects or checked for validity.
+* Fix an issue where `-[RLMSyncUser sessionForURL:]` could erronenously return a
+  non-nil value when passed in an invalid URL.
+* `SyncSession.Progress.fractionTransferred` now returns 1 if there are no
+  transferrable bytes.
 * Fix compilation issues with Xcode 8.3 beta 2.
 
 2.4.2 Release notes (2017-01-30)